Updated Fitness_center.dart tab one to properly fetch data from the api
o	When the user presses the Find Workouts button on the center of the screen, the variable exercises will be populated with exercise data by calling the getExercises() function.
o	The current implementation uses Navigator to route the user to a new page that will return a scaffold of a list of the workouts on a successful fetch.
	The goal is to eventually have this all natively happen within the Discover tab while providing the user with filters / selections to make.
	Currently the list view only provides the exercise name but changes will be made to provide more information.