Implemented FlipCard view for generated Workouts
The list of workouts now is presented in a flipcard view with the name of the exercise is on the front card alongside the various properties of the exercise, while details of how to perform the exercise are on the back of the card. Still early implementation but better than what was previously there.